requirements.txt
streamlit
google-generativeai
pandas
openpyxl
pymupdf
Pillow